Added new StatCraft Maven plugin to speed up build times and make builds simpler Tons of changes that I don't really remember

+++ src/main/java/wav/demon/StatCraft/Magic/MoveCode.java
@@ -28,7 +28,7 @@ public enum MoveCode {
         return code;
     }
 
-    public static MoveCode fromCode(short code) {
+    public static MoveCode fromCode(byte code) {
         for (MoveCode moveCode : values()) {
             if (code == moveCode.getCode())
                 return moveCode;
